requirements.txt
opencv-python
progressbar2
numpy~=2.0.1